requirements.txt
django~=1.9.6
djangorestframework==3.3.3
beautifulsoup4==4.4.1
bleach==1.4.3
django-debug-toolbar==1.4
django-crispy-forms==1.6.0
django-htmlmin==0.9.1
django-mptt==0.8.4
fake-factory~=0.5.7
html5lib<0.99999999,>=0.999
markdown==2.6.6
path.py==8.2.1
pillow==3.2.0
psycopg2==2.6.1
pytz
requests<3
unidecode
django-ckeditor
oss2
